LayoutTests/fast/events/spatial-navigation/snav-iframe-with-offscreen-focusable-element-expected.txt

a

e
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "1"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "1"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "1"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "2"
PASS document.activeElement.getAttribute("id") is "end"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "2"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "2"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "2"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "1"
PASS document.activeElement.getAttribute("id") is "start"
